flink/order-processor/requirements.txt
prometheus-client==0.19.0
requests==2.31.0
kafka-python==2.0.2